Reorganize remaining mixed unit tests
