+++ src/polaris-api-csharpTests/Methods/Cancellation/PatronSearchTests.cs
@@ -1,0 +1,38 @@
+using System;
+using System.Net.Http;
+using System.Threading;
+using System.Threading.Tasks;
+using Clc.Polaris.Api;
+using Clc.Polaris.Api.Models;
+using Clc.Polaris.Api.Tests;
+using Clc.Polaris.Api.Tests.TestInfrastructure;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+namespace Clc.Polaris.Api.Tests.Methods.Cancellation
+{
+    [TestClass]
+    [DoNotParallelize]
+    [UnitCategory]
+    public class PatronSearchTests : PapiClientTestBase
+    {
+        [TestMethod]
+        public async Task PatronSearchAsync_CancellationDuringProtectedTokenAuthentication_PropagatesCancellation()
+        {
+            var handler = new ProtectedTokenHttpMessageHandler();
+            var client = CreateProtectedClient(handler);
+            using var cancellationTokenSource = new CancellationTokenSource(TimeSpan.FromMilliseconds(10));
+
+            try
+            {
+                await client.PatronSearchAsync("name=Smith", cancellationToken: cancellationTokenSource.Token);
+                Assert.Fail("Expected cancellation to propagate.");
+            }
+            catch (OperationCanceledException)
+            {
+            }
+
+            Assert.AreEqual(1, handler.AuthenticationRequestCount);
+            Assert.AreEqual(0, handler.NonAuthenticationRequestCount);
+        }
+    }
+}
