+++ src/polaris-api-csharpTests/Methods/RequestShape/PatronUpdateTests.cs
@@ -1,0 +1,38 @@
+using System;
+using System.Net;
+using System.Net.Http;
+using System.Threading.Tasks;
+using Clc.Polaris.Api;
+using Clc.Polaris.Api.Models;
+using Clc.Polaris.Api.Tests;
+using Clc.Polaris.Api.Tests.TestInfrastructure;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+namespace Clc.Polaris.Api.Tests.Methods.RequestShape
+{
+    [TestClass]
+    [UnitCategory]
+    public class PatronUpdateTests : PapiClientTestBase
+    {
+        [TestMethod]
+        public async Task PatronUpdate_RequestShape_IsStable()
+        {
+            var handler = new CapturingHttpMessageHandler("{\"PAPIErrorCode\":0}");
+            var client = CreateClient(handler);
+
+            var response = await client.PatronUpdateAsync("AB C/+#?=", new PatronUpdateParams { EmailAddress = "patron@example.test" }, "1234", ignoresa: true);
+
+            Assert.IsNotNull(response);
+            Assert.AreEqual(HttpMethod.Put, handler.LastRequest!.Method);
+            var encodedBarcode = WebUtility.UrlEncode("AB C/+#?=");
+            StringAssert.Contains(handler.LastRequest.RequestUri!.AbsolutePath, $"/public/v1/1033/100/1/patron/{encodedBarcode}");
+            var query = ParseQuery(handler.LastRequest.RequestUri.Query);
+            Assert.AreEqual("True", query["ignoresa"]);
+            Assert.IsNotNull(handler.LastRequest.Content);
+            Assert.IsTrue(handler.LastRequest.Headers.Contains("PolarisDate"));
+            Assert.IsTrue(handler.LastRequest.Headers.Contains("Authorization"));
+            Assert.IsNotNull(handler.LastRequestContent);
+            StringAssert.Contains(handler.LastRequestContent, "patron@example.test");
+        }
+    }
+}
