+++ src/polaris-api-csharpTests/Models/PapiRestRequestPathClassificationTests.cs
@@ -1,0 +1,28 @@
+using System;
+using System.Threading.Tasks;
+using Clc.Polaris.Api;
+using Clc.Polaris.Api.Models;
+using Clc.Polaris.Api.Tests;
+using Clc.Polaris.Api.Tests.TestInfrastructure;
+using Clc.Rest.Models;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+namespace Clc.Polaris.Api.Tests.Models
+{
+    [TestClass]
+    [UnitCategory]
+    public class PapiRestRequestPathClassificationTests : PapiClientTestBase
+    {
+        [TestMethod]
+        public void PapiRestRequest_PathClassification_IsStable()
+        {
+            var publicRequest = new PapiRestRequest("/public/foo");
+            Assert.IsTrue(publicRequest.IsPublicMethod);
+            Assert.IsFalse(publicRequest.IsProtectedMethod);
+
+            var protectedRequest = new PapiRestRequest("/protected/foo");
+            Assert.IsFalse(protectedRequest.IsPublicMethod);
+            Assert.IsTrue(protectedRequest.IsProtectedMethod);
+        }
+    }
+}
