+++ src/polaris-api-csharpTests/PapiClient/PreformatRestRequestIdempotencyTests.cs
@@ -1,0 +1,43 @@
+using System;
+using System.Linq;
+using System.Net.Http;
+using System.Threading.Tasks;
+using Clc.Polaris.Api;
+using Clc.Polaris.Api.Models;
+using Clc.Polaris.Api.Tests;
+using Clc.Polaris.Api.Tests.TestInfrastructure;
+using Clc.Rest.Models;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+namespace Clc.Polaris.Api.Tests.PapiClient
+{
+    [TestClass]
+    [UnitCategory]
+    public class PreformatRestRequestIdempotencyTests : PapiClientTestBase
+    {
+        [TestMethod]
+        public void PreformatRestRequest_WhenCalledTwice_ReplacesPapiHeadersWithoutDuplicatingOrChangingBody()
+        {
+            var client = CreateClient();
+            var body = new { Value = "unchanged" };
+            var request = new PapiRestRequest(HttpMethod.Put, "/public/v1/1033/100/1/patron/ABC123", "1234", body);
+            request.Headers.Add("X-Custom", "keep");
+            request.QueryParameters.Add("ignoresa", false);
+
+            var first = (PapiRestRequest)client.PreformatRestRequest(request);
+            var firstHeaderCount = first.Headers.Count;
+            var second = (PapiRestRequest)client.PreformatRestRequest(first);
+
+            Assert.AreSame(first, second);
+            Assert.AreEqual(firstHeaderCount, second.Headers.Count);
+            Assert.AreEqual("keep", second.Headers["X-Custom"]);
+            Assert.AreSame(body, second.Body);
+            Assert.AreEqual(1, second.Headers.Keys.Count(key => key == "PolarisDate"));
+            Assert.AreEqual(1, second.Headers.Keys.Count(key => key == "Authorization"));
+            var date = second.Headers["PolarisDate"];
+            var expectedUri = "https://example.test/PAPIService/REST/public/v1/1033/100/1/patron/ABC123?ignoresa=False";
+            var expectedHash = PapiSignature.ComputeHash("access-key", "PUT", expectedUri, date, "1234");
+            Assert.AreEqual($"PWS access-id:{expectedHash}", second.Headers["Authorization"]);
+        }
+    }
+}
