+++ src/polaris-api-csharpTests/PapiClient/ProtectedTokenCacheTests.cs
@@ -1,0 +1,35 @@
+using System;
+using System.Net.Http;
+using System.Threading.Tasks;
+using Clc.Polaris.Api;
+using Clc.Polaris.Api.Models;
+using Clc.Polaris.Api.Tests;
+using Clc.Polaris.Api.Tests.TestInfrastructure;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+namespace Clc.Polaris.Api.Tests.PapiClient
+{
+    [TestClass]
+    [DoNotParallelize]
+    [UnitCategory]
+    public class ProtectedTokenCacheTests : PapiClientTestBase
+    {
+        [TestMethod]
+        public void BuildProtectedTokenCacheKey_UsesCredentialFingerprintWithoutRawSecrets()
+        {
+            var client = CreateProtectedClient(new ProtectedTokenHttpMessageHandler());
+            client.AccessKey = "raw-access-key-secret";
+            client.StaffOverrideAccount = CreateStaffUser(password: "correct-1");
+
+            var cacheKey = BuildCacheKey(client);
+
+            Assert.IsNotNull(cacheKey);
+            Assert.IsFalse(cacheKey!.Contains(client.StaffOverrideAccount.Password, StringComparison.Ordinal), "Cache key exposed secret material.");
+            Assert.IsFalse(cacheKey.Contains(client.AccessKey, StringComparison.Ordinal), "Cache key exposed secret material.");
+            StringAssert.Contains(cacheKey, client.Hostname.Trim());
+            StringAssert.Contains(cacheKey, client.AccessID.Trim());
+            StringAssert.Contains(cacheKey, client.StaffOverrideAccount.Domain.Trim());
+            StringAssert.Contains(cacheKey, client.StaffOverrideAccount.Username.Trim());
+        }
+    }
+}
